requirements.txt
discord.py
datetime
tk
asyncio
requests
python-dotenv
